Removed enumeration Type from class Row: added enumerations PolyhedronKind and RowKind; they are merged into a single object using new class Row::Type. Rows now rememberif they have or not the \epsilon coefficient.
Constructors for Matrix now have a boolean flag indicating whether or not
the matrix has the \epsilon coefficient. The flag is remembered in order
to correctly set the type of new rows.

Added support for creating constraints that are strict inequalities and
generators that are closure points. Added support for adding these two new
kind of objects to constraint systems and generator systems (with
automatic detection and resizing if needed).

Constraints, generators and the respective systems space-dimensions
now depend on the Polyhedron kind of the corresponding rows.