Added template <class Box> C_Polyhedron::C_Polyhedron(const Box& box) and template <class Box> NNC_Polyhedron::NNC_Polyhedron(const Box& box). Added the declaration and documentation and sketched the implementation of template <class Box> Polyhedron::Polyhedron(Topology topol, const Box& box): these constructor allow to builds a polyhedron out of a generic, interval-based bounding box.
Documentation of Polyhedron::shrink_bounding_box(Box& box) revised: we
now promise to communicate rational numbers using fractions in
canonical form.  The fraction n/d is in canonical form if and only if
n and d have no common factors and d is positive, 0/1 being the unique
representation for zero.