In the Prolog interface, exceptions now always thrown if input values (other than lists) are invalid.
Tests for these new exceptions added to the pl_check.pl test file.

Corrected bug in Prolog interface in the case when a list was expected
but a non-list was input.

The interface needs to be changed so that it throws an exception for
non-lists (instead of just Prolog failure) when a list is expected;
this job has been added to the TODO file.