Return a bool from the update_congruences and simplify methods to indicate whether the grid is empty.  Pass this return on in the minimize and add_and_minimize methods which call simplify.  Assert an equal number of columns in source1 and source2 in add_and_minimize\5 (congruence to generator).
In update_congruences and update_generators set the grid empty
according to the minimize return.

In both simplifys use the row number instead of the row to test if the
original row and pivot must be swapped.  In the congruence simplify
when adding the top row, search for a valid modulus and set the
diagonal (and modulus) to the result of the search.