A BD_Shape representation now also carries a dbm matrix of Boolean values, indicating whether or not the corresponding dbm entry is redundant. The new matrix is meaningful only if the BDS_Status test     marked_shortest_path_reduced() returns true, otherwise it is meaningless. Note that a dbm can be shortest-path reduced only if it is also shortest-path closed. The redundancy information is established by shortest_path_reduction_assign() and is lost as soon as we modify the representation of the BD_Shape. That is, currently no effort is done to keep it up-to-date. Moreover, when we reset the shortest-path closure flag, then the reduction flag is reset too.
Added private helper methods
    add_dbm_constraint()
that allow for the addition of a difference bound constraint without
encoding it into a Constraint object. The method is particularly used
in the computation of affine images.

The methods computing affine images and preimages rewritten (almost)
from scratch. (The old implementations, were not behaving correctly
when the denominator was negative or when considering generalized images).

Added a lot of const qualifications; avoided some dbm's coefficient copies,
preferring the use of references instead.