There are now three flavours of grid widening.
The `generator_widening_assign()' is as commited previously
(using generators) and file `widening2.cc' tests this.

The `congruence_widening_assign()' is a renaming of the old
`widening_assign()' (using congruences) and `widening1.cc' tests this.

The new `widening_assign()' uses either of the above two widenings
depending on which of the systems describing the grid are
up-to-date. The new test file `widening3.cc' tests this widening.

There are, similarly, three flavours of a
`limited_*extrapolation_assign()' operator.

The documentation in `definitions.dox' on the grid widening
has been brought up to date and improved.