In ppl_interface_generator_common.m4, changed the macro name set_schema_strings to replace_all_patterns.
The same name change in C/ppl_interface_generator_c_dat.m4
and Prolog/ppl_interface_generator_prolog_dat.m4 where
this macro is defined.

In the Prolog interface, the and_embed and and_project
combined into one schema as already done in the C interface.

In response to a FIXME comment in the C interface code,
the concatenate_assign is now included in the schema corresponding
to the binop pattern. This change has only been done in the
Prolog interface.