Corrected a bug introduced for formatting.
Revised the use of generator to grid_generator to match changes
due to merge from the trunk.

Removed the use of alt_box replacement for the grid "get_covering_box"
and "get_bounding_box", by providing the code explicitly.

Now the "alt_" replacement is not used anymore and this
option for defining replacements will be removed later
when other improvements here are made..