Revised the way we handle compound domains (is the powerset and product domains).
Where compound domains have Polyhedron as a component, the system
creates two domains, one where Polyhedron is prefixed by C and the
other where it is prefixed by NNC. Note that this change has been made
so as _not_ to affect the simple domain Polyhedron.

As requested, both the user name and the C++ name have the topology C
and NNC inserted just before the string "Polyhedron". Note though this
still assumes that every component = Polyhedron has the same
topology. Automatically generating different topologies for these
components will be difficult and not really practical or useful.

Simplified the way class name patterns are replaced by the actual
name. The class patterns are now replaced as part of the main list of
replacements.