Improved the processing of the replacement lists for a pattern.
Added 2 new application independent helper macros:
 - m4_nargs (definition taken from m4 documentation)
 - m4_arg (finds the n+1'st argument in a list of arguments -
where n is the 1st argument)

The alternative replacement is now also defined via an alternative list
of replacements (the number and order must match the usual list of
replacements for the given class). This change was triggered by the
wish to have a "friend" pattern (see below).

Introduced a new pattern "friend" and use in in the Prolog interface
to use instead of the 2nd "class" in
ppl_new_@CLASS@_from_@INTOPOLOGY@@CLASS@
ie.,
ppl_new_@CLASS@_from_@INTOPOLOGY@@FRIEND@.
That is, a friend is a class that can be used to construct an element
in the current class. Every class is declared to be a friend of itself.
The intention is to add to more "friend"s for each class so
that, for instance, Polyhedron will be a friend of BD_Shape.
Note that the friend pattern is only used in the Prolog interface at
the moment.