Added 3 new application-independent macros to the common.m4 m4_echo_quoted, m4_echo_unquoted, m4_arg_counter:
m4_echo_quoted and m4_echo_unquoted correspond to echo1 and echo2
in the m4 documentation. Neither used yet but I plan to use
the m4_echo_unquoted to expand terms.

m4_arg_counter takes a string and a sequence of arguments
and either expands to the first index of the argument sequence
that matches the string or to the empty string.
m4_arg_counter is now used instead of m4_check_if_class_in_group
which is removed.

In common_dat.m4 added a dummy predicate for widening
Polyhedra_Powerset classes and in prolog_icc_code.m4 added
unused schematic code for developing the widening for powersets
(this is prefixed by a "FIXME" and should be ignored).