Removed a duplicate macro definition, improved some comments, reordered some macros.
Added macros to make the components of a "simple" open or direct
product cplusplus class available; these are now defined as:
m4_cplusplus_class`'i_j where "i" is the class counter and "j" = 1 or 2
depending on whether it is the 1st or second argument.
By "simple", we mean that the product construct is not nested.