Error in the swap code corrected.
Tests for space_dimension and affine_dimension added.

Now, for check_noisy, a list of all successful tests is output.
For check_all and check_noisy, a list of all failed tests is output.
For check_noisy, the constraint and generator (or similar) systems
of the failing object are output.