In ppl/src:
The grid can now be built from boxes which may have different finite
values for the upper and lower bounds or have only the upper or the
lower bound bounded or have open finite bounds. Before when this
happened an exception was thrown. Now we compute an approximation.

Also, now the Grid domain requires the complexity as
an argument for shrink_bounding_box so as to be compatible with the
Polyhedron_shrink_bounding_box. This argument is not used.

The tests/Grid/boundingbox1.cc and tests/Grid/boundingbox2.cc have
been revised to match.