In the ppl/interfaces:
The @BOX@ replacement is now only used for building from a box
(bounding_box or covering_box).

The "get" or "shrink" bounding_box or covering_box" procedures, do not
use any pattern. As the Grid has the same aruments as the Polyhedron
for shrink_bounding_box, we can combine all the classes for this
procedure into one schema. On the other hand, although the covering
box is only available at present for the Grid domain, it is provided
as a generic schema.

The property "is_topologically_closed" is now enabled for BD_Shape and
Octagons Prolog interfaces.
Similarly the "topological_closure_assign" procedure is now enabled for
BD_Shape and Octagons Prolog interfaces.

Generic tests for all the properties using schema
ppl_@CLASS@_@HAS_PROPERTY@ have been added.  Generic tests for
ppl_@CLASS@_@SIMPLIFY@ have been added (for the domains under
consideration, this is just the ."topological_closure_assign"
procedure).

The test data for the generic tests has again been reorganised and
restructured making it easier to extend and maintain.