Added tests for comparing the class objects.
General improvements to the test system including:

Changing the test_data identifiers to be "test0n".
Using a list to give all the required tests and the member predicate
to select the tests.
Generating code for building the test object for any given test id
and use this instead of coding it in the actual tests.
The space_dimension and affine_dimension are now defined by an
auxiliary predicate for each test id.