Revised the grid relation_with() methods for congruences and constraints so that saturates() is included in the result whenever the congruence or constraint is an equality and every point of the grid satisfies the equality.
Added grid relation_with() methods for a generator so that a point or
closure point has the same relation with the grid as the corresponding
grid point, and a ray or line has the same relation with the grid as
a grid line.

Expected results revised and new tests for the grid relation_with()
methods for a generator added to relations1.cc.