Expanded the number of predicates generated for classes such as Octagonal_Shape<...> and Pointset_Powerset<...>.
Fixed a bug in the m4 code for generating ppl_prolog.icc
when building a class from a box.
Note that the bug only was apparent for classes with a templatic argument
such as Octagonal_Shape<...>.

In the m4 code for generating ppl_predicate_check.pl
tests modified to allow for the larger set of classes.