Enable make predicate_check_test for Ciao. To do this, the ":- include ppl_predicate_check_common." directive has to be in the generated file ppl_predicate_check.pl
For this reason, the corresponding directives in swi_predicate_check.pl
and SICStus/sp_predicate_check.pl have been removed.

For XSB, the include directive is not supported as is and instead
a "#include" must be used. Changes made for XSB to handle this.

Warning messages given by the Ciao system when compiling
the generated file ppl_predicate_check.pl need further investigation
but the test in Ciao succeeds.