Corrected a missing "+" in interfaces/ppl_interface_generator_common_procedure_generators.m4 which caused some interface predicates not to be generated.
Added add_constraint_and_minimize() and add_constraint_and_minimize()
to the Partially_Reduced_Product domain.

In the interfaces, the interface procedures for
   ppl_Grid_generalized_image_lhs_rhs/5 and
   ppl_Grid_generalized_preimage_lhs_rhs/5
   ppl_Grid_generalized_image/6 and
   ppl_Grid_generalized_preimage/6
   have been renamed as
   ppl_Grid_generalized_image_lhs_rhs_with_congruence/5
   ppl_Grid_generalized_preimage_lhs_rhs_with_congruence/5
   ppl_Grid_generalized_image_with_congruence/6
   ppl_Grid_generalized_preimage_with_congruence/6
   so as to allow for /4 and /5, resp., versions

(This change also reported in "News").

Tidied and commented the procedure list in
interfaces/ppl_interface_generator_common_procedure_generators.m4
so that where not "+all" domains are implemented in the interfaces,
comments explain further.

A few remaining procedures that can be enabled for products
and/or powersets in the interfaces have been enabled.

Added two "todo" procedures to the OCaml list of procedure generators.