Added hash_code() to the pointset_powerset and partially_reduced_product domains. Enable this in the Java interface for all the domains. Also enable in the Java interface the string() method for all the domains, so as to include the product domains.
Added methods for external/total_memory_in_bytes for all the domains
and to all the interfaces.
Existing tests for size() for prolog and java now use the MEMBYTES
pattern instead so as to test all the replacements for MEMBYTES.

These with the size() method (for the pointset_powerset domains)
are all generated from a single "pattern" "MEMBYTES", defined in
interfaces/ppl_interface_generator_common_dat.m4.

Added methods for external/total_memory_in_bytes for all the domains
and to all the interfaces.

Existing tests for size() for prolog and java now use the MEMBYTES
pattern instead so as to test all its replacements.

These with the size() method (for the pointset_powerset domains)
are all generated from a single "pattern" "MEMBYTES", defined in
interfaces/ppl_interface_generator_common_dat.m4.