Enabled the Pointset_Powerset domain to have a Rational_Box as the disjunct and the Product domains to have a Rational_Box as one of its component domains in all the interfaces.
To allow for a number of missing methods in the Box domain that
cannot therefore be provided in any product domain that has the
Box as a component, defined two generic macros in
ppl_interface_generator_common.m4 that intersect sequences of
comma-separated strings: one that intersects two sequences and
one that intersects three sequences.

Also now support a new notation in a procedure string -
e.g., "?box" means don't include this procedure for any
product that has Box as one of its components.
This option may be removed in the future and its use
avoided by using the sequence intersection macros described
above.