The code for intersecting sequences of comma separated strings improved so that the same basic code is used for intersecting both 2 sequences and 3 sequences.
The last changes introduced a bug for some products (such as
Grid x BD_Shape<...> products. This is now fixed, but the change has
a "fixme" since this indicates that the way the replacements for all
patterns is handled needs improving and that this local change
should be made more general.