Changes mainly affect poly_difference_assign().
Added difference_assign() to be a clone of poly_difference_assign().
Added a couple of tests and improved an existing test.

Replaced all lines:
// FIXME: this is buggy when PS is not an abstraction of NNC_Polyhedron.
by
// This code is only used when PS is an abstraction of NNC_Polyhedron.
since the Grid domains are specialised separately.

Corrected a bug in poly_difference_assign() for Grids where some
precision was wrongly lost (when a remainder was negative and had
been assumed to be positive).