Corrected two bugs in the revised code for constructing a grid from a polyhedron with complexity class:
The space dimension of the generator system must be set to be
that of the polyhedron.

After minimizing the polyhedron (in the case that the complexity is
ANY_COMPLEXITY), we must check if the polyhedron is empty, and if
it is return the empty grid.

Added a new test file for checking this constructor, moved
existing tests for it that were in grid3.cc to the new test file,
and added a few extra tests to check the use of complexity classes.