Added to the Partially_Reduced_Product<> class in the semantic GD constructors from other classes a complexity parameter.
Tests for this have been added as needed. Also some comments
wrt the tests for the product domain have been corrected.

In the Grid, Box, BD_Shape and Octagonal_Shape added
a complexity class parameter to the copy constructor
and to the constructor from the same class but with a different
numeric class (where applicable).

Test files added where no obvious test for the copy constructor
could be found. For the grid tests, the test file copyconstruct1.cc
has been renamed for uniformity with other tests and the other
classes.