The method topology() made public.
The code for upper_bound_assign_if_exact() added to the C interface.
The predicates for the same methods in the Prolog interface
now no longer need the topology  (for the Polyhedron class) in the name.

The schematic procedure name for these procedures also no longer has
the topology and therefore the use of these in the other interfaces
(Java and OCaml) has also been adapted.

Some updating of the Prolog interface files for other recent changes.