Fixed a problem in the code generation for linear_partition and approximate_partition.
This problem occurred in the Prolog interfaces when the list of
instantiations did not include
Pointset_Powerset<NNC_Polyhedron> and/or Pointset_Powerset<Grid>.

Note that each of the Prolog systems generates headers/declarations
for every instance of every active schema in the procedure list.
Therefore if the code is not required because the set of instantiations
does not include Pointset_Powerset<NNC_Polyhedron>
(resp., Pointset_Powerset<Grid>)
we have also to ensure that the header/declaration is not generated.