The core ppl_interface_generator_common.m4 file has been completely reorganised, macros improved and comments revised and added. Note that alternative patterns and replacements now have a simple _a_ or _b_ extension instead of _alt_ or _cppx_. To change the case of a word, use "!" instead of "L" or "U" before the pattern.
The macros in the core file are organised into groups and a comment
summarises the call structure between the groups.
Each macro is labelled to indicate if it is used locally,
just in this file but by a macro in another group,
or if it is public and may be used by other code generating m4 files.
Suggestions for further improving the macro descriptions are welcome.

The complementary file  ppl_interface_generator_common.m4_dat file
that contains the common data and some macros for generating
the basic class names, the patterns and their replacements
is improved and some redundant things removed.
Commenting is not improved but much of the structure here is simpler
and existing comments provide a lot of information already.

The changes to these files has meant that changes to all the m4 files
in the interfaces directory was required.

These changes seem to make the m4 generation slightly more efficient -
but the improvement appears to be less that 10% and disappointing.