Changes to deal with //FIXME(0.10.1): the following is a bug! in src/Pointset_Powerset.templates.hh
As Grids are non-convex, the constructor for a powerset of nnc
polyhedra from a powerset of grids has been implemented separately
from the generic case. A test added to show that there was a bug which
is now fixed.

The "//FIXME(0.10.1): ..." has been removed but a new FIXME added
indicating this code will break if other non-convex or non-linear
domains are added