Added a workaround to a space efficiency bug affecting MIP_Problem. The baseline patch has been contributed by Marco Poletti.
MIP_Problem now stores the input constraints as a sequence of Constraint*.
In this sequence, we distinguish between owned constraints (whose resources
are under control of the current MIP_Problem object) and inherited constraints
(whose resources are shared with, and owned by, another MIP_Problem object).
When starting a branch-and-bound computation looking for an integral
feasible/optimal solution, we now inherit (rather than copy) the input
constraints.

The workaround greatly improves space/time efficiency on (mixed) integral
problems, also providing a (simple and reasonable) partial solution to

    https://www.cs.unipr.it/mantis/view.php?id=62

The solution is still partial: a complete solution requires the implementation
of a backtracking mechanism for incrementally added constraints.