Replaced the algorithm used for sorting Linear_System and Bit_Matrix.
The new algorithm, based on STL's std::sort and std::unique, is applied
to a vector of rows' indices and implements indirect sort using function
objects for the comparisons and swap operations.
Apart from a (usually minor) decrease in the number of comparisons and
swaps perfomed, the new algorithm should be better wrt the consumption
of stack space, by getting rid of ad-hoc sort/unique functions, as well as
the special-purpose iterator class Linear_System::With_Bit_Matrix_iterator.