Set INLINE_SOURCES = NO when producing LaTeX documentation.
This change is a workaround to a pdftex + hyperref issue that may result
in the following error
   \pdfendlink ended up in different nesting level than \pdfstartlink
when using some versions of Doxygen (e.g., 1.7.3).

According to http://tug.org/errors.html the error is triggered when a
hyperlink is split across page boundaries and it can be generated even
by valid LaTeX code. Disabling source inlining is a way to greatly decrease
its likelihood.