Do not call is_sorted() in method remove_row_no_check().
Method is_sorted() performs an assertion check regarding sortedness flag,
but in method remove_row_no_check() we are in the middle of changes and
the `sorted' flag is not (yet) up-to-date.
This fixes the crash in test02() of tests/Polyhedron/removespacedims2.cc.

While at it, corrected "not-pending" as "non-pending" in several places.