Implemented the quick adjacency check in the conversion procedure.
The test allows for significant performce improvements on certain classes
of polyhedra (e.g., hypercubes) in that it saves many saturation subset checks.

Added preprocessing flags to allow for enabling/disabling the new check
(as well as the old quick non-adjacency check). Added new testcase
  tests/Polyhedron/hypercubes.cc
that can be used for a performance comparison.

The efficiency benefits can also be appreciated when running
the ppl_lcdd demo testsuite.