Avoid a memory leak due to the global list of dirty temporaries.
The memory leak was not causing a real problem, in that the list of dirty
temporaries is meant to survive up to the end of the process.
However, it could cause a problem if/when the library is extended to become
multi-thread safe, since in such a scenario it is meaningful to have
a list of dirty temporaries for each thread (to be released on thread exit).