The ppl_lpsol demo program now supports thread safety.
If multiple input files are given as input, each of them is processed
by a separate thread. Note: the file reading phase, which used glpk,
is performed by the master thread (since glpk is not thread safe).
The maximum number of threads is fixed at compilation time (MAX_THREADS = 8).
In multi-threaded mode the meaning of option -o is modified to specify
the output *directory*.

Commit partially based on previous work by Maxim Gaina.