Improved thread-safety for the mechanism for abandoning expensive computations.
Added helper make_interruptible() to allow for the creation of threads
executing PPL-related computations that can be interrupted.
Note: it is up to the library user to do what is needed in order to
propagate the exceptions from the (interrupted) worker thread to the
(interrupting) master thread. See tests/Polyhedron/threadsafe2.cc
for an example.

Note: the mechanism is only available in the C++ interface.