New test07() shows a bug in the filtering of minimized NNC generators.
The bug should only affect output operations; it is not going to affect
the result of methods such as NNC_Polyhedron::is_topologically_closed().

The bug has been discovered while working on an alternative representation
and implementation for NNC polyhedra, requiring no epsilon dimension.
The conversion algorithm for this new representation will be described
in the undergraduate thesis of Anna Becchi (work in progress).